feat: move algorithm enum to dedicated module
